+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## 2023-12-27
+
+### Added
+
+- Added the ability to favorite builds. Favorited builds will be available on your profile.
+
 ## 2023-12-26
 
 ### Changed